requirements.txt
Flask
strawberry-graphql[debug-server]
pytest
pytest-asyncio